LayoutTests/platform/mac/accessibility/aria-image-emits-object-replacement-expected.txt


This tests ensures that if an ARIA image is used it will emit an object replacement character in a range for string operation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text1.textMarkerRangeLength(textMarkerRange) is 5
Object string for range: a ￼ b
PASS text1.textMarkerRangeLength(textMarkerRange) is 5
Object string for range: c ￼ d
PASS successfullyParsed is true

TEST COMPLETE
